+++ jpostman-annotations/src/test/java/io/jpostman/annotations/runtime/JPostmanAnnotationResponseFilterRegressionTest.java
@@ -157,12 +157,8 @@ public class JPostmanAnnotationResponseFilterRegressionTest {
 
 			profileExecutions++;
 			int execution = profileExecutions;
-			String body = "{"
-					+ "\"id\":" + execution + ","
-					+ "\"firstName\":\"Emily-" + execution + "\","
-					+ "\"lastName\":\"Johnson-" + execution + "\","
-					+ "\"gender\":\"female-" + execution + "\""
-					+ "}";
+			String body = "{" + "\"id\":" + execution + "," + "\"firstName\":\"Emily-" + execution + "\","
+					+ "\"lastName\":\"Johnson-" + execution + "\"," + "\"gender\":\"female-" + execution + "\"" + "}";
 			return okExecutor(body);
 		}
 	}
@@ -218,12 +214,8 @@ public class JPostmanAnnotationResponseFilterRegressionTest {
 
 			profileExecutions++;
 			int execution = profileExecutions;
-			String body = "{"
-					+ "\"id\":" + execution + ","
-					+ "\"firstName\":\"Emily-" + execution + "\","
-					+ "\"lastName\":\"Johnson-" + execution + "\","
-					+ "\"gender\":\"female-" + execution + "\""
-					+ "}";
+			String body = "{" + "\"id\":" + execution + "," + "\"firstName\":\"Emily-" + execution + "\","
+					+ "\"lastName\":\"Johnson-" + execution + "\"," + "\"gender\":\"female-" + execution + "\"" + "}";
 			return okExecutor(body);
 		}
 	}